Moved search functionality to separate class
unwrapped the functionality so it may be expanded on easier
