+++ requirements.txt
@@ -2,3 +2,4 @@ click==7.1.2
 Flask==1.1.2
 qrcode==6.1
 waitress==1.4.4
+regex
